rename 2.1D to 2.2D
